+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.4.0] - 2025-02-25
 
 ### Added
+- server-side authentication for the status-reviews page ([216](https://github.com/boarlabsxyz/Kavoon/pull/216))
 - generation of sitemap.xml for Google indexing ([#218](https://github.com/boarlabsxyz/Kavoon/pull/218))
 
 ### Fixed